backend/requirements.txt
# Core Django Dependencies
Django==4.2.7

# API Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-migration-testcase==1.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7
pycryptodome==3.19.0

# Blockchain
web3==6.15.1
eth-utils==2.3.0
eth-account==0.10.0

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage & IPFS
ipfshttpclient==0.8.0a2
Pillow==10.1.0
django-storages==1.14.2

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development Tools
python-decouple==3.8
django-debug-toolbar==4.2.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# ML Dependencies
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
